% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soma.R
\name{create_protdata_from_soma}
\alias{create_protdata_from_soma}
\title{A protData contructor for SomaScan Data}
\usage{
create_protdata_from_soma(adat, condition = NULL, filter = TRUE)
}
\arguments{
\item{adat}{}
}
\description{
A protData contructor for SomaScan Data
}
