% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-prot_data.R
\name{detect_intensity_cols}
\alias{detect_intensity_cols}
\title{Detect_intensity_cols}
\usage{
detect_intensity_cols(df)
}
\arguments{
\item{df}{}
}
\value{
list of column numbers that re numeric (intensity)
}
\description{
Detect_intensity_cols
}
