% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{do_limma_by_condition}
\alias{do_limma_by_condition}
\title{Perform DEA using the condition labels of the protdata object}
\usage{
do_limma_by_condition(object, condition, control_group, treatment_group)
}
\arguments{
\item{treatment_group}{}
}
\description{
Perform DEA using the condition labels of the protdata object
}
