% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{do_t_test,ProtData-method}
\alias{do_t_test,ProtData-method}
\title{Title}
\usage{
\S4method{do_t_test}{ProtData}(object, treatment_samples, control_samples)
}
\arguments{
\item{ProtData}{}
}
\description{
Title
}
