% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{enrich_go}
\alias{enrich_go}
\title{Title}
\usage{
enrich_go(
  gene_id,
  all_gene_vector,
  enrich_pvalue = 1,
  org = org.Hs.eg.db::org.Hs.eg.db
)
}
\arguments{
\item{enrich_pvalue}{}
}
\description{
Title
}
