% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{enrich_kegg}
\alias{enrich_kegg}
\title{Title}
\usage{
enrich_kegg(
  gene_id,
  all_gene_vector,
  enrich_pvalue = 1,
  org = org.Hs.eg.db::org.Hs.eg.db,
  organism = "hsa"
)
}
\arguments{
\item{enrich_pvalue}{}
}
\description{
Title
}
