% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{enrich_pathways}
\alias{enrich_pathways}
\title{Title}
\usage{
enrich_pathways(
  DE,
  lfc_threshold = 1,
  fdr_threshold = 0.01,
  enrich_pvalue = 0.05
)
}
\arguments{
\item{enrich_pvalue}{}
}
\description{
Title
}
