% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{filter_features}
\alias{filter_features}
\title{Helper function to filter out sparse proteins}
\usage{
filter_features(DT_limma, control_samples, treatment_samples, alpha)
}
\description{
Helper function to filter out sparse proteins
}
