% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{gse_kegg}
\alias{gse_kegg}
\title{Title}
\usage{
gse_kegg(
  gene_list,
  enrich_pvalue = 1,
  org = org.Hs.eg.db::org.Hs.eg.db,
  organism = "hsa"
)
}
\arguments{
\item{organism}{}
}
\description{
Title
}
