% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{plot_pg_counts}
\alias{plot_pg_counts}
\title{Title}
\usage{
plot_pg_counts(PD, by_condition = F)
}
\arguments{
\item{PD}{A prot_data object}

\item{by_condition}{True if you want to group by condition}
}
\description{
Title
}
