% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{plot_pg_intensities}
\alias{plot_pg_intensities}
\title{Title}
\usage{
plot_pg_intensities(PD)
}
\arguments{
\item{PD}{prot_data object}
}
\value{
ggplot boxplot of intensities for each sample
}
\description{
Title
}
