% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_outliers.R
\name{remove_outliers,ProtData-method}
\alias{remove_outliers,ProtData-method}
\title{Title}
\usage{
\S4method{remove_outliers}{ProtData}(object, sds = 3)
}
\arguments{
\item{ProtData}{}
}
\description{
Title
}
