% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prot_data.R
\name{create_protdata}
\alias{create_protdata}
\title{Create a ProtData Object}
\usage{
create_protdata(data, condition = NULL, method = "Unknown")
}
\arguments{
\item{data}{A data frame containing proteomics data (proteins are rows, samples are columns).}

\item{condition}{A data frame containing conditions of the samples. Rownames should match colnames of data. Optional.}

\item{method}{A character string describing the method used for generating the data. Optional.}

\item{prot_meta}{A data frame containing metadata for the proteins (rows). Optional.}
}
\value{
An instance of the ProtData class.
}
\description{
This function creates an instance of the ProtData class.
}
