% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{do_limma}
\alias{do_limma}
\title{do_limma() takes in a protData object, the column names of the treatment group and control group.
The function first removes features (proteins) that are found in less than 50\% of control and treatment
samples. Returns a dataframe with the protein metadata, intensity values for the control and treatment groups,
the logfc and pvalue/adjusted pvalue.}
\usage{
do_limma(object, treatment_samples, control_samples, meta_col)
}
\arguments{
\item{object}{}
}
\description{
do_limma() takes in a protData object, the column names of the treatment group and control group.
The function first removes features (proteins) that are found in less than 50\% of control and treatment
samples. Returns a dataframe with the protein metadata, intensity values for the control and treatment groups,
the logfc and pvalue/adjusted pvalue.
}
