% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Title}
\usage{
plot_volcano(
  DT.original,
  label_col = NULL,
  lfc_threshold = 1,
  fdr_threshold = 0.01,
  labelgene = NULL
)
}
\arguments{
\item{labelgene}{}
}
\description{
Title
}
